requirements.txt
python-telegram-bot==13.7
pandas==1.5.3
numpy==1.24.3
openpyxl==3.0.9
fpdf2==2.5.5
pytz==2021.3
python-dotenv==1.0.0